Fix errors reported only on PHP 8.2

+++ src/MassiveAction.php
@@ -414,8 +414,7 @@ class MassiveAction
             $this->timer->start();
             $this->fields_to_remove_when_reload[] = 'timer';
 
-            /** @var number $max_time */
-            $max_time = get_cfg_var("max_execution_time");
+            $max_time = (int) get_cfg_var("max_execution_time");
             $max_time = ($max_time == 0) ? 60 : $max_time;
 
             $this->timeout_delay                  = ($max_time - 3);
